backend/remote-state/gcs: Rename "gcloud" to "gcs" for backwards compatibility.
