+++ backend/remote-state/gcs/backend_state.go
@@ -1,4 +1,4 @@
-package gcloud
+package gcs
 
 import (
 	"fmt"
