+++ backend/remote-state/gcs/client.go
@@ -1,4 +1,4 @@
-package gcloud
+package gcs
 
 import (
 	"encoding/json"
